icarusStream/requirements.txt
keyboard==0.13.5
opencv_python_headless==4.6.0.66
PyYAML==6.0
